Generic parameters for nested types should only be explicitly written
Basically, when we ask for the type name, we ask for it as a literal, with generics set to false. Even if we have generics set to false, we still want to get our declaring type's generic parameters, but if we are ourselves AND we have generics set to false, we want to ensure that we don't suffix our name with our OWN generic parameter(s).